hehe

+++ bot/modules/cb_handler.py
@@ -7,6 +7,10 @@ from bot.helpers.database.database import fetch_media_details, change_video_type
 
 @Client.on_callback_query(filters.regex(pattern="helpmsg"))
 async def help_cb(c: Client, cb: CallbackQuery):
+    user_id = cb.data.split("_")[1]
+    if int(user_id) != cb.from_user.id:
+        await cb.answer(lang.NOT_AUTH_CB)
+        return
     buttons = await help_buttons()
     await c.edit_message_text(
         chat_id=cb.message.chat.id,
@@ -17,6 +21,10 @@ async def help_cb(c: Client, cb: CallbackQuery):
 
 @Client.on_callback_query(filters.regex(pattern="botinfo"))
 async def botinfo_cb(c: Client, cb: CallbackQuery):
+    user_id = cb.data.split("_")[1]
+    if int(user_id) != cb.from_user.id:
+        await cb.answer(lang.NOT_AUTH_CB)
+        return
     buttons = await main_menu_buttons()
     await c.edit_message_text(
         chat_id=cb.message.chat.id,
@@ -27,6 +35,10 @@ async def botinfo_cb(c: Client, cb: CallbackQuery):
 
 @Client.on_callback_query(filters.regex(pattern="uploadhelp"))
 async def upload_files_help_cb(c: Client, cb: CallbackQuery):
+    user_id = cb.data.split("_")[1]
+    if int(user_id) != cb.from_user.id:
+        await cb.answer(lang.NOT_AUTH_CB)
+        return
     buttons = await upload_helper_buttons()
     await c.edit_message_text(
         chat_id=cb.message.chat.id,
@@ -37,6 +49,10 @@ async def upload_files_help_cb(c: Client, cb: CallbackQuery):
 
 @Client.on_callback_query(filters.regex(pattern="changevideotype"))
 async def change_video_type_cb(c: Client, cb: CallbackQuery):
+    user_id = cb.data.split("_")[1]
+    if int(user_id) != cb.from_user.id:
+        await cb.answer(lang.NOT_AUTH_CB)
+        return
     user_id = cb.message.reply_to_message.from_user.id
     user_name = cb.message.reply_to_message.from_user.first_name
     video_type, photo_type = await fetch_media_details(user_id)
@@ -57,6 +73,10 @@ async def change_video_type_cb(c: Client, cb: CallbackQuery):
 
 @Client.on_callback_query(filters.regex(pattern="changephototype"))
 async def change_photo_type_cb(c: Client, cb: CallbackQuery):
+    user_id = cb.data.split("_")[1]
+    if int(user_id) != cb.from_user.id:
+        await cb.answer(lang.NOT_AUTH_CB)
+        return
     user_id = cb.message.reply_to_message.from_user.id
     user_name = cb.message.reply_to_message.from_user.first_name
     video_type, photo_type = await fetch_media_details(user_id)
@@ -77,6 +97,10 @@ async def change_photo_type_cb(c: Client, cb: CallbackQuery):
 
 @Client.on_callback_query(filters.regex(pattern="close"))
 async def close_cb(c: Client, cb: CallbackQuery):
+    user_id = cb.data.split("_")[1]
+    if int(user_id) != cb.from_user.id:
+        await cb.answer(lang.NOT_AUTH_CB)
+        return
     await c.delete_messages(
         chat_id=cb.message.chat.id,
         message_ids=cb.message.message_id
